# Conflicts:
#	.idea/vcs.xml
#	app/src/main/java/org/androidtown/actionbar/MainActivity.java

+++ app/src/main/java/org/androidtown/actionbar/Fragment01.java
@@ -12,7 +12,8 @@ public class Fragment01 extends Fragment {
 	private static final String ARG_PARAM1 = "온도";
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState ) {
-		int num = getArguments().getInt(ARG_PARAM1);
+		//double num = getArguments().getInt(ARG_PARAM1);
+		double num = getArguments().getDouble(ARG_PARAM1);
 //		int num=40;
 		View view = inflater.inflate(R.layout.frag01, container, false);
 		TextView tv = (TextView) view.findViewById(R.id.text01);
